So I don't need to use a database for giveaway.
