+++ src/events/client/eventUserRemove.js
@@ -6,18 +6,18 @@ export default {
     name: Events.GuildScheduledEventUserRemove,
     once: false,
     execute: async (guildScheduledEvent, user) => {
-        try {
-            await UserForScheduledEvents.deleteOne({
-                userId: user.id,
-                guildScheduledEventId: guildScheduledEvent.id,
-            });
-            console.log(
-                chalk.redBright(
-                    `${user.tag} has left the event ${guildScheduledEvent.name}`,
-                ),
-            );
-        } catch (error) {
-            console.error(error);
-        }
+        // try {
+        //     await UserForScheduledEvents.deleteOne({
+        //         userId: user.id,
+        //         guildScheduledEventId: guildScheduledEvent.id,
+        //     });
+        //     console.log(
+        //         chalk.redBright(
+        //             `${user.tag} has left the event ${guildScheduledEvent.name}`,
+        //         ),
+        //     );
+        // } catch (error) {
+        //     console.error(error);
+        // }
     },
 };
